feat: GameObjRaw methods as a prototype
