+++ src/ecs/components/init/paused.ts
@@ -1,0 +1,11 @@
+import type { GameObj } from "../../../types";
+
+const PausedComp = {
+    add(this: GameObj) {
+        this.paused = true;
+    },
+};
+
+export function paused() {
+    return PausedComp;
+}
